#4 Moved everything regarding mesos master into own class
