+++ elasticsearch-cloud-mesos/src/main/java/org/elasticsearch/cloud/mesos/MesosMasterStateService.java
@@ -1,0 +1,12 @@
+package org.elasticsearch.cloud.mesos;
+
+import org.elasticsearch.discovery.mesos.MesosUnicastHostsProvider;
+
+import java.util.List;
+
+/**
+ * A singleton to manage all the tasks from mesos.
+ */
+public interface MesosMasterStateService {
+    List<String> getNodeIpsAndPorts(MesosUnicastHostsProvider mesosUnicastHostsProvider);
+}
