--- elasticsearch-cloud-mesos/src/main/java/org/elasticsearch/cloud/mesos/MesosTaskService.java
@@ -1,7 +1,0 @@
-package org.elasticsearch.cloud.mesos;
-
-/**
- * A singleton to manage all the tasks from mesos.
- */
-public interface MesosTaskService {
-}
